# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Signature Agent Bridge API",
    "version": "0.1.0",
    "description": "OpenAI-compatible Chat Completions and unified /v1/bridge orchestration extensions. Copyright 2026 Signature Management Consultants SLU. Author @ancongui.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8766"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/v1/bridge/status": {
      "get": {
        "operationId": "get_v1_bridge_status",
        "summary": "GET /v1/bridge/status",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs": {
      "get": {
        "operationId": "get_v1_bridge_jobs",
        "summary": "GET /v1/bridge/jobs",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}": {
      "get": {
        "operationId": "get_v1_bridge_jobs_id",
        "summary": "GET /v1/bridge/jobs/{id}",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}/audit": {
      "get": {
        "operationId": "get_v1_bridge_jobs_id_audit",
        "summary": "GET /v1/bridge/jobs/{id}/audit",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "through",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Fix this to the first page's through value for an immutable export."
          },
          {
            "name": "includePayloads",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "default": "false"
            },
            "description": "Include complete tool inputs/results. Pages stop around 4 MB, always retaining at least one whole entry."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}/tools": {
      "get": {
        "operationId": "get_v1_bridge_jobs_id_tools",
        "summary": "GET /v1/bridge/jobs/{id}/tools",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "filesOnly",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "default": "false"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolPage"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}/tools/{tool}": {
      "get": {
        "operationId": "get_v1_bridge_jobs_id_tools_tool",
        "summary": "GET /v1/bridge/jobs/{id}/tools/{tool}",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tool",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolObservation"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}/attempts": {
      "get": {
        "operationId": "get_v1_bridge_jobs_id_attempts",
        "summary": "GET /v1/bridge/jobs/{id}/attempts",
        "description": "Requires read authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}/pause": {
      "post": {
        "operationId": "post_v1_bridge_jobs_id_pause",
        "summary": "POST /v1/bridge/jobs/{id}/pause",
        "description": "Requires control authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "control",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/bridge/jobs/{id}/resume": {
      "post": {
        "operationId": "post_v1_bridge_jobs_id_resume",
        "summary": "POST /v1/bridge/jobs/{id}/resume",
        "description": "Requires control authorization. See api.md for lifecycle semantics, errors, and response fields.",
        "tags": ["Jobs"],
        "x-required-scope": "control",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; error.code identifies the reason.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sab_ token"
      }
    },
    "schemas": {
...
      "ToolPage": {
        "type": "object",
        "properties": {
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolObservation"
            }
          },
          "nextCursor": {
            "type": ["integer", "null"]
          }
        },
        "required": ["tools", "nextCursor"],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "param": {
                "type": "null"
              },
              "requestId": {
                "type": "string"
              },
              "job_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": ["code", "message", "type", "param"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      }
    }
  }
}